etl_airflow_pandas/requirements.txt
pandas
apache-airflow==2.6.1
prometheus-client
pytest